+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvTimeSeriesSourceTest.groovy
@@ -28,8 +28,7 @@ import java.time.ZoneId
 class CsvTimeSeriesSourceTest extends Specification implements CsvTestDataMeta {
 	def "The csv time series source is able to provide a valid time series mapping from files"() {
 		given:
-		def coordinateSource = Mock(IdCoordinateSource)
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
+		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy())
 		def expectedMapping = [
 			new TimeSeriesMapping.Entry(UUID.fromString("58167015-d760-4f90-8109-f2ebd94cda91"), UUID.fromString("b86e95b0-e579-4a80-a534-37c7a470a409"), UUID.fromString("9185b8c1-86ba-4a16-8dea-5ac898e8caa5")),
 			new TimeSeriesMapping.Entry(UUID.fromString("9a9ebfda-dc26-4a40-b9ca-25cd42f6cc3f"), UUID.fromString("c7ebcc6c-55fc-479b-aa6b-6fa82ccac6b8"), UUID.fromString("3fbfaa97-cff4-46d4-95ba-a95665e87c26")),
@@ -45,125 +44,12 @@ class CsvTimeSeriesSourceTest extends Specification implements CsvTestDataMeta {
 		expectedMapping.stream().allMatch { mappingEntries.contains(it) }
 	}
 
-	def "The csv time series source is able to build a single WeatherValue from field to value mapping"() {
-		given:
-		def defaultCoordinate = GeoUtils.DEFAULT_GEOMETRY_FACTORY.createPoint(new Coordinate(7.4116482, 51.4843281))
-		def coordinateSource = Mock(IdCoordinateSource)
-		coordinateSource.getCoordinate(5) >> Optional.of(defaultCoordinate)
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
-		def fieldToValues = [
-			"uuid": "71a79f59-eebf-40c1-8358-ba7414077d57",
-			"time": "2020-10-16T12:40:42Z",
-			"coordinate": "5",
-			"directirradiation": "1.234",
-			"diffuseirradiation": "5.678",
-			"temperature": "9.1011",
-			"windvelocity": "12.1314",
-			"winddirection": "15.1617"
-		]
-		def expectedValue = new TimeBasedValue(
-				UUID.fromString("71a79f59-eebf-40c1-8358-ba7414077d57"),
-				TimeUtil.withDefaults.toZonedDateTime("2020-10-16 12:40:42"),
-				new WeatherValue(
-				defaultCoordinate,
-				new IrradiationValue(
-				Quantities.getQuantity(1.234, IRRADIATION),
-				Quantities.getQuantity(5.678, IRRADIATION)
-				),
-				new TemperatureValue(
-				Quantities.getQuantity(9.1011, TEMPERATURE)
-				),
-				new WindValue(
-				Quantities.getQuantity(12.1314, WIND_DIRECTION),
-				Quantities.getQuantity(15.1617, WIND_VELOCITY)
-				)
-				)
-				)
-
-		when:
-		def actual = source.buildWeatherValue(fieldToValues)
-
-		then:
-		actual.present
-		actual.get() == expectedValue
-	}
-
-	def "The csv time series source returns no WeatherValue, if the coordinate field is empty"() {
-		given:
-		def defaultCoordinate = GeoUtils.DEFAULT_GEOMETRY_FACTORY.createPoint(new Coordinate(7.4116482, 51.4843281))
-		def coordinateSource = Mock(IdCoordinateSource)
-		coordinateSource.getCoordinate(5) >> defaultCoordinate
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
-		def fieldToValues = [
-			"uuid": "71a79f59-eebf-40c1-8358-ba7414077d57",
-			"time": "2020-10-16T12:40:42Z",
-			"coordinate": "",
-			"directirradiation": "1.234",
-			"diffuseirradiation": "5.678",
-			"temperature": "9.1011",
-			"windvelocity": "12.1314",
-			"winddirection": "15.1617"
-		]
-
-		when:
-		def actual = source.buildWeatherValue(fieldToValues)
-
-		then:
-		!actual.present
-	}
-
-	def "The csv time series source returns no WeatherValue, if the coordinate field is missing"() {
-		given:
-		def defaultCoordinate = GeoUtils.DEFAULT_GEOMETRY_FACTORY.createPoint(new Coordinate(7.4116482, 51.4843281))
-		def coordinateSource = Mock(IdCoordinateSource)
-		coordinateSource.getCoordinate(5) >> defaultCoordinate
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
-		def fieldToValues = [
-			"uuid": "71a79f59-eebf-40c1-8358-ba7414077d57",
-			"time": "2020-10-16T12:40:42Z",
-			"directirradiation": "1.234",
-			"diffuseirradiation": "5.678",
-			"temperature": "9.1011",
-			"windvelocity": "12.1314",
-			"winddirection": "15.1617"
-		]
-
-		when:
-		def actual = source.buildWeatherValue(fieldToValues)
-
-		then:
-		!actual.present
-	}
-
-	def "The csv time series source returns no WeatherValue, if the coordinate cannot be obtained"() {
-		given:
-		def coordinateSource = Mock(IdCoordinateSource)
-		coordinateSource.getCoordinate(6) >> Optional.empty()
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
-		def fieldToValues = [
-			"uuid": "71a79f59-eebf-40c1-8358-ba7414077d57",
-			"time": "2020-10-16T12:40:42Z",
-			"coordinate": "6",
-			"directirradiation": "1.234",
-			"diffuseirradiation": "5.678",
-			"temperature": "9.1011",
-			"windvelocity": "12.1314",
-			"winddirection": "15.1617"
-		]
-
-		when:
-		def actual = source.buildWeatherValue(fieldToValues)
-
-		then:
-		!actual.present
-	}
-
 	def "The csv time series source is able to build time based values from simple data"() {
 		given:
 		def defaultCoordinate = GeoUtils.DEFAULT_GEOMETRY_FACTORY.createPoint(new Coordinate(7.4116482, 51.4843281))
 		def coordinateSource = Mock(IdCoordinateSource)
 		coordinateSource.getCoordinate(5) >> defaultCoordinate
-		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy(), coordinateSource)
+		def source = new CsvTimeSeriesSource(";", timeSeriesFolderPath, new FileNamingStrategy())
 		def factory = new TimeBasedSimpleValueFactory(EnergyPriceValue)
 		def time = TimeUtil.withDefaults.toZonedDateTime("2019-01-01 00:00:00")
 		def timeUtil = new TimeUtil(ZoneId.of("UTC"), Locale.GERMANY, "yyyy-MM-dd'T'HH:mm:ss[.S[S][S]]'Z'")
